Fix file path unit tests on windows - The pain never ends (:
